feat(core): moved to Symbol data type for mediator metadata
