+++ packages/core/lib/messages/constants.ts
@@ -1,10 +1,2 @@
-export const MESSAGE_METADATA = '__mediatorMessage__';
-
-export const SCOPE_OPTIONS_METADATA = '__mediatorScopeRequest__';
-
-export const GLOBAL_REQUEST_PIPELINE_METADATA = '__mediatorGlobalRequestPipeline__';
-export const REQUEST_PIPELINE_METADATA = '__mediatorRequestPipeline__';
-export const REQUEST_HANDLER_METADATA = '__mediatorRequestHandler__';
-
-export const GLOBAL_EVENT_HANDLER_METADATA = '__mediatorGlobalEventHandler__';
-export const EVENT_HANDLER_METADATA = '__mediatorEventHandler__';
+export const MESSAGE_METADATA = Symbol('mediatorMessage');
+export const SCOPE_OPTIONS_METADATA = Symbol('mediatorScopeOptions');
