+++ packages/core/lib/messages/event/constants.ts
@@ -1,0 +1,2 @@
+export const GLOBAL_EVENT_HANDLER_METADATA = Symbol('mediatorGlobalEventHandler');
+export const EVENT_HANDLER_METADATA = Symbol('mediatorEventHandler');
