+++ packages/core/lib/messages/event/decorators/event-handler.decorator.spec.ts
@@ -1,10 +1,11 @@
 import 'reflect-metadata';
-import { EVENT_HANDLER_METADATA, MESSAGE_METADATA, SCOPE_OPTIONS_METADATA } from '../../constants';
+import { MESSAGE_METADATA, SCOPE_OPTIONS_METADATA } from '../../constants';
 import { MessageTypeInterferenceException } from '../../exceptions/message-type-interference.exception';
 import { ScopeOptions } from '../../interfaces';
 import { IMessageMetadata } from '../../interfaces/message-metadata.interface';
 import { MessageTypes } from '../../message-types.enum';
 import { TestEvent, TestEventHandler } from '../../messages.mocks';
+import { EVENT_HANDLER_METADATA } from '../constants';
 import { EventHandler } from './event-handler.decorator';
 
 class AnotherEvent {}
