+++ packages/core/lib/messages/request/constants.ts
@@ -1,0 +1,3 @@
+export const GLOBAL_REQUEST_PIPELINE_METADATA = Symbol('mediatorGlobalRequestPipeline');
+export const REQUEST_PIPELINE_METADATA = Symbol('mediatorRequestPipeline');
+export const REQUEST_HANDLER_METADATA = Symbol('mediatorRequestHandler');
