+++ packages/core/lib/messages/request/decorators/request-handler.decorator.ts
@@ -1,9 +1,10 @@
 import { ClassConstructor } from '../../../utils/class-constructor.interface';
-import { MESSAGE_METADATA, REQUEST_HANDLER_METADATA, SCOPE_OPTIONS_METADATA } from '../../constants';
+import { MESSAGE_METADATA, SCOPE_OPTIONS_METADATA } from '../../constants';
 import { MessageTypeInterferenceException } from '../../exceptions/message-type-interference.exception';
 import { IMessageMetadata } from '../../interfaces/message-metadata.interface';
 import { ScopeOptions } from '../../interfaces/scope.options';
 import { MessageTypes } from '../../message-types.enum';
+import { REQUEST_HANDLER_METADATA } from '../constants';
 import { IRequest } from '../interfaces/request.interface';
 
 type RequestType = ClassConstructor<IRequest>;
