+++ packages/core/lib/messages/request/decorators/request-pipeline.decorator.ts
@@ -1,7 +1,8 @@
 import { ClassConstructor } from '../../../utils/class-constructor.interface';
-import { REQUEST_PIPELINE_METADATA, SCOPE_OPTIONS_METADATA } from '../../constants';
+import { SCOPE_OPTIONS_METADATA } from '../../constants';
 import { ScopeOptions } from '../../interfaces/scope.options';
 import { IRequest } from '../interfaces/request.interface';
+import { REQUEST_PIPELINE_METADATA } from '../constants';
 
 type RequestType = ClassConstructor<IRequest>;
