+++ packages/core/lib/providers-manager/messages/events/adapters/global-events-handlers.adapter.ts
@@ -1,6 +1,6 @@
 import { ClassConstructor } from '../../../../utils/class-constructor.interface';
 import { IEvent, IEventHandler, MessageTypes } from '../../../../messages';
-import { GLOBAL_EVENT_HANDLER_METADATA } from '../../../../messages/constants';
+import { GLOBAL_EVENT_HANDLER_METADATA } from '../../../../messages/event/constants';
 import { IProviderTypeAdapter } from '../../../ports/provider-type-adapter.port';
 import { IEventsProvidersSchema } from '../interfaces/events-providers-schema.interface';
